Add $TRAVIS_REPO_SLUG test for deployment
Master branch pushes to cloned repos were trying to upload, and finding
they could not because the AWS secret credentials were only for
`metaeducation/ren-c`.

This didn't usually happen since usually pushes are branches.  But it
was noticed by creating a fresh enlistment and trying to build its
master on Travis... seeing all the builds fail because it didn't know
not to try to upload.

Hence we leverage the environment variable $TRAVIS_REPO_SLUG for this
particular case.  (It may not be common, but it's the first one most
people would see if they tried their own Travis builds of the repo.)