Initial implementation of continuation/delegation
This is a preliminary demonstration of working with a mode in which
function executions can be revisited on multiple passes.  The
heap-allocated "Rebol frame" (REBFRM*) remains chained into the logical
stack of running functions while the function defers, so it can store
information in its cells regarding what to do once the deferred
execution has been performed by the evaluator.

When this approach is applied uniformly as the method of invoking
evaluator recursions, it becomes a "continuation-passing" style.  This
is desirable for several reasons--such as being able to recurse the
interpreter a number of times bounded by physical memory instead of
the more scarcely allocated C stacks.  But at a mechnical level, it
means that there is a way to check when no more recursions can be
performed (e.g. a memory allocation fails on making a stack frame)
whereas there is no clean or standards-compliant way to tell if adding
a regular stack frame will fail.

Continuations know they are on their first run if their frame's spare
cell is an END marker.  This cell can then be used as an out-of-band
place to store a state telling the native what continuation mode
it is in.  While natives could also use local cells for their state,
this is of the most benefit to dispatchers--which are trying to run
frames that are generic, and have nowhere to store state information.

When a plain continuation signal is used, the native will be called
back with the output value for further processing--interpreting its
position in its processing from the frame state.  But there is another
mode called "delegation", which assumes that the output does not
need further processing.  Beyond convenience, this could also be
helpful for cases where a dispatcher cannot be run due to damage to
the frame (such as how ENCLOSE will "steal" frames)

Convenience macros wrapping `return` with contiuation arguments are
made as CONTINUE() and DELEGATE().