Rearrange %c-eval.c to solve some skipped initializations
The evaluator is rather goto-happy, and that is somewhat by design--
many of the gotos are in service of capturing performance points and
optimizations in ways that inlining could not accomplish.  (To the
extent it's not by design, it's just because figuring out the best
way to lay out the evaluator is a question that evolves organically
as the features it offers are explored...but the assertions help guide
various restructurings to try and see where to make it better as
time permits for refactoring.)

C compilers (and especially C++ compilers) will warn you when you
use a goto that takes you to points in the code where a variable
skips its initialization.  Not all of them do so when the variable
does not get used after the jumped-to point...especially some older
compilers.  But in order to appease older compilers and keep the code
compiling there, this rearranges some of the gotos.  It seems to help
show some kind of direction in better ordering of the code.