Addition of GENERATOR and YIELDER
With "stackless" evaluation, what actually happens is that recursions
into the interpreter are accomplished with continuations...that are
represented via heap data structures.  This transforms stack overflow
errors into memory exhaustion errors--which can be detected when they
happen.

Another benefit of representing the interpreter stack with heap
structures is that the stack can be saved and restored, giving the
ability to jump into and out of code stacks.  This allows for the
implementation of generators.

This is an experiment with trying to implement this with GENERATOR
(though it cannot work with any constructs on the stack that have
not been rewritten to a stackless approach).  So only simple cases
can work at the moment:

    >> g: generator [yield 1 | yield 2]

    >> y
    == 1

    >> y
    == 1  ; jumped back in at the right execution point

    >> y
    ; null

Rather than force through a `make generator` step, this approach
intends to let the generator be used off the cuff.  An example
might be:

    >> array/initial 10 generator [
        while [true] [yield 1 yield 2 yield 3]
    ]
    == [1 2 3 1 2 3 1 2 3 1]

Those wishing to have a "generator maker" with parameterization
that is consistent across yields can build that explicitly:

    gmaker: function [i] [
        return generator [
           yield i
           yield i + 10
       ]
    ]

    gen: gmaker 10

    print [gen]
    ; expected output: 10

    print [gen]
    ; expected output: 20

While this approach may be less efficient when making many
generators, the more flexible approach is being explored before
trying to decide on a shape of optimizing it.

There is another form of generator which is parameterized, where
each invocation receives a parameter...yet still has the resuming:

    y: yielder [x] [
        yield x + 1
        yield x + 2
    ]

    >> y 10
    == 11

    >> y 20
    == 22

Implementing this correctly will ultimately mean that any global
state (such as the mold buffer) is marked and properly resumed.
For the moment, this just accounts for the data stack...by
preserving the portion of the stack above the yielder...also
adjusting the DSP indices at each point.