Expose Is_Throwing() as a non-debug feature
Ren-C does not have any particular datatype for a "throw", because
a throw necessarily would have to encompass an ANY-VALUE!...which would
be (by definition) larger than the frame's output cell.  So throwing
was a protocol understood by all calls into the evaluator as a boolean
that had to be checked as a result of that call.

So it was not seen as necessary to be able to query the "thrown state"
of the evaluator.  Anyone who was interested had been responsible for
paying attention to that flag, and reacting appropriately.

With the migration to stackless processing, each native that wants to
invoke the interpreter has to delegate to it...and gets called again
with a value to check in D_OUT.  But if there is a throw during that
evaluation, the dispatcher is not on the stack and has nowhere to put
the thrown state.

This speculatively implements the somewhat unappealing notion of making
asking about the thrown state a legal operation.  The debug build
still has some protections against ignoring the thrown status (e.g.
D_OUT will appear corrupt during a thrown state) but better methods
could be used to ensure that the Is_Throwing() was checked even if
no throw occurred (by corrupting D_OUT somehow in the "ordinary" case)