Make usermode function dispatches stackless
The existing model by which usermode functions ran through the same
dispatcher as a native meshes well with the concept of being able
to have a return value that encapsulates the desire to execute a
continuation.  Since most usermode dispatchers are little more than
asking to execute the block stored with the function, the simpler
ones were trivial.

But unlike a native which can adapt its definition to add more
arguments to use to track the state of how many times it has been
called, the dispatchers are supposed to work generically on user
defined frames where they do not control the number of arguments.
In order to give these dispatchers something to work with, the
"spare" cell (which is used during argument fulfillment and other
times) is given a usefully initialized state to END.  This makes it
possible to use as meaningful state that can start a sequence of
continuation phases.

This can also be used by natives--so the IF stackless variant is
changed from having a local for tracking re-entry to using the spare.
It also promotes the stackless variant to being the used version,
as well as modifying EITHER to follow the pattern.