Change feeds to nominally use dynamic allocation
Moving frames from using C stack storage to heap storage was needed
to begin the changes for "stackless" processing.  Yet feeds were also
starting to need dynamic allocation.

This changes the default to assume that dynamic allocation is used.
Whether a frame is responsible for freeing a feed when it is dropped
is set as an evaluator flag, so that regardless of whether a raised
error or otherwise happens the feed still gets freed.

A debug mode is included for catching the tick on which a feed was
allocated if it is not freed.