Eliminate REB_P_RETURN parameter class
Initial design of the definitional return mechanism used a special
parameter class that was identified during the frame filling in order
to preload it with a special ACTION! variant of RETURN which was
bound to the executing frame.

This led to some amount of inefficiency, in the sense that natives did
not actually use these RETURN cells (they only used the typeset in the
frame key for listing return types, and for type checking in the
debug build).  It also raised semantic questions about how a feature
that was intended to be something users could do themselves was doing
magic that function compositions couldn't obviously do.

But the system firmed up with the role of dispatchers and natives, and
the position of the return on functions with ACTION_FLAG_HAS_RETURN
standardized to the first argument slot (despite that slot being
what looked like the "first argument" prior to better locals handling).
This makes it fairly clear that no parameter class should be needed...
common code for the dispatchers can fill the slot.

(This may even permit phases to re-engineer the value to make the
return checks apply to the specific phase that's running.)