Reclaim NODE_FLAG_TRANSIENT
The "Transient" feature was used to distiguish data stack cells from
other stack cells.  It was in order to indicate that those cells
could not be targets of an evaluation...and hence no code could write
into them directly with values that had pointers that would potentially
need to be managed.

This allowed cells with unmanaged bindings to be put into the data
stack.  How much value this really had is questionable--a test for
if a pointer is in the data stack could be done with bounds checking of
the range of the data stack if it was really considered important.

More valuable in the near term is the ability to distinguish between
a REBFRM* and a REBVAL* with a strict-aliasing safe test.  Having a
bit in the first byte is one easy way to do it (not necessarily the
ideal one for long term).  But this may help make continuations more
efficient in general.