Fix use of EVAL_MASK_DEFAULT over FEED_MASK_DEFAULT
The bit patterns for EVAL and feeds are different.  In GET_XXX_FLAG()
and SET_XXX_FLAG() errors are reduced by making those macros do
token pasting.  But that doesn't catch some cases.

Since it was caught by an experimental use of NODE_FLAG_4 to
indicate REBFRM* "nodes", that is kept in this commit for the moment.