Get rid of EVAL_FLAG_PUSH_TO_STACK
Was added for REDUCE2, which was solved another way (by being able to
actually ask for frames to run single eval steps stacklessly).  There
is perhaps theoretically a use for such a mode, but it's premature
optimization since you can't do any examination of the values as they
are being reduced.  Better to optimize the continuations themselves.