Eliminate continuation_type
In the first attempt to create continuations, they would label the
frame that was being processed non-recursively with the type of
the value that had triggered the evaluation...coming from the types
of branches in Do_Branch_Throws().

As the labels evolved, they did not look very obvious (e.g. REB_BLANK
was used to indicate a frame was spawned by the argument fulfillment
process).  And it seemed that there was a recurring pattern in the
after-handling that split the desires into a few categories of state
transitions when the continuation frame completed.

This picks up the slack with other aspects of frame state (including
new frame flags).  It's a step on the way to making the handling of
continuations behave in a more homogenous and systemic way, as the
needs of a stackless system continue to be addressed.