Add initial concept of a "frame executor"
With the elimination of "continuation types", certain frame flags and
state were used to decide what kind of dispatch would pick up when
a continuation completed.  These were ad hoc and switched on in the
emerging "master frame evaluation" loop--that would keep processing
whatever the top frame was.

This takes a small step toward putting a dispatcher in the frame that
indicates what will take over when the continuation is done.  This
allowed for the elimination of two frame flags--for starters.  But
since this is a very primitive idea, it is still double-checked.

Ultimately this aims to grow into a solution where a native dispatcher
could also count, and record itself as intending to be called back
or to be chained.