Re-enable BLANK!, QUOTED!, and SYM-XXX! branch types
Some branch types were disabled when continuations were changed to
be frame-based only, because they did not produce frames...only an
f->out output pointer.  This would be interpreted as the frame's
final answer, meaning that there'd be no callback into the requesting
continuation's state machine for it to post-process the output.

This makes a simple no-op frame that can handle the result in those
non-evaluative cases.  It also attempts to introduce a frame type
that can "double-evaluate" to do SYM-GROUP! evaluations, but it does
not work yet due to a problem with calling back non-ACTION!-based
frames.