Create Group_Executor() for stackless GROUP! handling
GROUP! presents an unusual case due to `1 + (comment "hi") 2` being
legal.  This required special handling different from the typical
Do_Any_Array() in the non-stackless evaluator, and so it means it needs
a different kind of continuation in the stackless one.

This takes a step toward getting the right behavior, and tries out the
idea of calling the frame functions "XXX_Executor()".