Explore adding output parameter to interpreted continues
This is a preservation of adding a parameter to let interpreted
delegations write somewhere other than f->out, e.g. a frame's spare
cell.  The idea did not pan out--due to the nature of RETURN in
Elider_Dispatcher().  There were notes on that fact, but it was not
in quite an obvious enough place to notice.

The idea could have uses in other places, but continuations writing
anywhere than their requesting frame's f->out raises questions.  This
leaves the questions in, but still keeps the added parameter for now,
with an assert that it matches `f->out`.  It's easier to delete than
to add in again, and the comments can serve as a warning if the ideas
involved get forgotten.