Rethink varlist conservation
This makes the reuse of varlist stubs more clear, with comments.

In the attempt to make the invariants more obvious, the difficulties
that the MATCH command has with FULFILL_ONLY frames was noticed.  This
led to a general realization about the need for some kind of invariant
identity for enumeration frames, which would not be conflated with
the identity of function call frames.

That is explored in subsequent commits--while this one simply breaks
the MATCH command.