Trampoline tolerance of calling non-top-frame executors
Before the trampoline was written, there was no reified form for
enumeration "frames" (e.g. the kind that a REDUCE or ANY or CASE might
use to lock a hold on a block/va_list and walk across it). So if
something like one of those natives wanted to take itself off the stack
to do an enumeration, where would it store this state...and in what form?

This commit implements a version of that answer: to just store it on
the frame stack, stacked deeper than the native.  But dispatchers for
natives are able to get called back into with the knowledge they are
not the top frame on the stack.  (This ability already existed, as a
REBFRM* parameter called `frame_` was always being passed, it just had
historically always corresponded to the top of the stack).

Making it possible for a native to be called with a frame that is not
the mechanical top of the stack adds a bit more nuance to the
trampoline's workings.  But it seems a viable way for a native wishing
to do evaluations to safely hold onto an enumeration state.

Pleasingly, this resurrects the previous idea that FRAME! identity
need only be reified for ACTION! frames (as opposed to inventing
some new type of FRAME! or FEED! abstraction).  This seems a very
good tradeoff vs. expecting natives to always be on the top of the
stack for their entry points.

Retrofits the stackless REDUCE2 implementation to the new convention.