Make YIELD preserve the state of output cell
Some natives use f->out as a kind of "working storage space" when
their work is done in phases.  If a YIELD happens across this state,
then the output cell of the yielder gets overwritten.  It's necessary
to preserve the value somewhere.

This preservation has to also keep the OUT_MARKED_STALE bit in sync,
and it needs to be able to preserve an END state--illegal in blocks.
That is a accomplished with a trick that uses a hidden identity (the
identity of the details array) to set up a dummy BLOCK! value that
means "END".