Renamings for clarity and fix GROUP! behavior
This makes the sequence of steps in the expression evaluation more
clear, and puts the responsibility of clearing the OUT_MARKED_STALE
bit on the end of the expression evaluation chain (Finished_Executor())

Because the Group_Executor() runs outside of the evaluation cycle, this
means it is no longer convenient for it to read the stale bit.  So
it uses its spare cell to target the evaluation, so it can determine
by virtue of being END at the end of a sub-evaluation that the GROUP!
vaporized, without conflating with risking clearance of the bit on
the sole `f->out` cell.