Add KEEP_STALE_BIT and rethink CASE predicates
Due to problems with the logic behind "leaking" stale bits (and comments
not being as clear as they probably could be), the OUT_MARKED_STALE
handling and asserts relating to it were messy or disabled in the first
attempts to do stackless evaluation.  It seemed like it may not be
worth it compared to performing evaluations into cells other than f->out
and using the END status to decide whether to copy the cell back.

That inefficiency undercuts the premise behind the stale bit; which
may come into play in particular when cell movement causes series to
be marked as managed (due to being in more than one place).  It seemed
to interfere with writing a clean version of stackless CASE, where
having access to the stale bit would solve some problems.

This goes ahead and makes an evaluator flag for not clearing the stale
bit in the Finished_Executor(), that may be optionally used in
constructs that want to manage it themselves.  The function dispatch
will still make sure no functions return stale marked values.

In addition, this works through a compromise on handling predicates in
CASE, which permits predicates of arity > 2, and allows them to run
without spawning an additional frame independent of the enumeration.
This was prohibited previously due to concerns that it would interfere
with DEFAULT, but the compromise is to switch to a default predicate
semantics of IDENTITY instead of DID.  This can be made to allow the
DEFAULT hack to still work, even if the predicate runs on it.