Make COUNT a state variable
It's not completely clear which variables in the parse executor are
going to need to be preserved across continuations, or when subparse
recursions would be able to cover up holding local state.

This goes ahead and puts the count for a repeated rule into a state
variable, which avoids putting too much thought into worrying about
`goto` over its initialization.  It also allows it to be set to an
"unreadable blank" state in order to catch usages when it was not
thought to be initialized properly.

Future optimizations may rethink this--but the unreadable state will
provide guidance on whether that's possible.