Reuse EVAL_FLAG_TO_END to carry meaning of ~PF_ONE_RULE
The PF_ONE_RULE flag was introduced as a "find flag", not realizing
that FIND_FLAG_XXX were sharing states with AM_XXX flags in R3-Alpha.

This takes question of whether one rule or many should be executed and
unifies that with the EVAL_FLAG_TO_END--already used for the same
meaning in the plain evaluator (e.g. "DO/NEXT" vs. plain DO).